% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_arrange_share_legend.R
\name{grid_arrange_share_legend}
\alias{grid_arrange_share_legend}
\title{Common legend for two or more `gg` grobs.}
\usage{
grid_arrange_share_legend(
  ...,
  nrow = 1,
  ncol = length(list(...)),
  position = c("bottom", "right")
)
}
\arguments{
\item{...}{Objects of class grob}

\item{nrow}{Number of rows to align figures in.}

\item{ncol}{Number of columns to align figures in.}

\item{position}{Position where the legend will be printed. "bottom" or
"right"}
}
\value{
A grob
}
\description{
Almost all parts of the function is borrowed from `vignette(package = "gridExtra", "arrangeGrob")`.
Refer to this vignettee for extra details.
}
\note{
Source: http://rpubs.com/sjackman/grid_arrange_shared_legend
}
\examples{
\dontrun{
# Example
mtmodified <- mtcars %>%
mutate_at(c("am", "vs"), as.factor)
mtmodified1 <- mtmodified %>% ggplot(aes(am, mpg, group = `am`, fill = `am`)) + geom_boxplot()
mtmodified2 <- mtmodified %>% ggplot(aes(vs, mpg, group = `vs`, fill = `vs`)) + geom_boxplot()
grid_arrange_share_legend(mtmodified1, mtmodified2,
                          nrow = 2, ncol = 1, position = "bottom")
}
}
