% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_lm_list.R
\name{hello}
\alias{hello}
\title{Hello name!}
\usage{
hello(x)
}
\arguments{
\item{x:}{A string}
}
\value{
The output from \code{\link{print}}
}
\description{
This is an example function named 'hello'
which greets any "name" input by printing "Hello name!".
}
\examples{
hello("Deependra")

}
