% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_genetics_tidy.R
\name{heritability_n_blues}
\alias{heritability_n_blues}
\title{Compute heritability and genotypic BLUEs for each selected grouping variable adjusting for covariate.}
\usage{
heritability_n_blues(df, grouping_var, grouping_items_sel, dependent_var,
  covariate = NULL, genotype_var, replication_var)
}
\arguments{
\item{df}{A dataframe or similar object of coercible class.}

\item{grouping_var}{Column name of the grouping variable such as location or environment of character or factor type.}

\item{grouping_items_sel}{A character vector of items which are subset of grouping variable selected for obtaining heritability.}

\item{dependent_var}{Column name of response variable whose heritability estimates are being obtained through model.}

\item{covariate}{Column name of numeric variable. Adjusted values of dependent variable for this covariate will be used in obtaining estimates.}

\item{replication_var}{Column name of variable representing replicated records of genotypes/treatments.}

\item{genotype}{Column name of variable representing genotypes/treatments.}
}
\value{
A list
}
\description{
Compute heritability and genotypic BLUEs for each selected grouping variable adjusting for covariate.
}
\note{
Adapted from 'meta-r' application package
}
\examples{

library(lme4)
library(emmeans)
require(tidyverse)
# load data
df <- data(sample_rcbd) # na = ".", while reading the csv
df <- df \%>\%
  mutate_at(c("Genotype", "Rep"), as.factor)
# check for factor levels
df \%>\%
  select_if(funs(is.factor(.)|is.character(.))) \%>\%
  map_int(~length(unique(.x)))
# calculate genetic correlation for locations in absence of other grouping factors while adjusting for covariate
heritability_n_blues(df, grouping_items_sel = c("Buenos Aires", "Beijin", "Karaj"),
                     grouping_var = Location,
                     genotype = Genotype, replication_var = Rep,
                     dependent_var = YLD, covariate = TGW)

}
