% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary_longstargazer.R
\name{long_stargazer}
\alias{long_stargazer}
\title{Long stargazer}
\usage{
long_stargazer(...)
}
\arguments{
\item{...}{Arguments to be passed regularly in `stargazer::stargazer`}
}
\value{
TeX code
}
\description{
Easy and useful wrapper around stargazer that instantly embeds given
latex table in a longtable environment.
}
\note{
Must use this function with float=FALSE in stargazer

While using stargazer in a chunk use results = "asis" in option.
}
\examples{
\dontrun{
# some nice defaults for longstargazer
}

}
