% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mean_differences.R
\name{mean_differences_emmeans}
\alias{mean_differences_emmeans}
\title{Tidies up a list of lme (or lm) models to obtain table of mean comparison}
\usage{
mean_differences_emmeans(model_list, treatment_factor_list)
}
\arguments{
\item{model_list}{A list of model objects (\code{\link{lme}}, \code{\link{lmer}} or \code{\link{lm}})}

\item{treatment_factor_list}{A list of treatment factor(s) for which contrasts are specified in the model}
}
\value{
A tbl
}
\description{
This function relies extensively on `emmeans` package (yes, it works for lm too)
}
\examples{
\dontrun{
# main effects and interaction effects means for two factors using emmeans tidier
mean_differences_emmeans(model_list = mt_named_list, treatment_factor_list = mt_treatment_factor_list)

# plot emmeans
plot(emmeans::emmeans(mt_lm1, "am"), comparison = T)

# plot emmeans p-values
emmeans::pwpp(emmeans::emmeans(mt_lm1, "am")) %>% plot()
}

}
