% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary_kabletype.R
\name{p_annotator}
\alias{p_annotator}
\title{Regression table with summary annotations (like stargazer) using knitr}
\usage{
p_annotator(model_list)
}
\arguments{
\item{model_list}{A list of model objects generated by lm_list or custom generated list}
}
\value{
A tbl
}
\description{
This function accepts a list of models, for multiple response variables,
and even when fitted with different model terms
this is similar to `stargazer::stargazer` but more friendly to tweak
It accepts a list of model (i.e. `lm`, `lme`, etc.) with (suitably some) common terms
It returns a tibble with annotations -- the stars correspond to the level of significance,
the standard error are shown in parenthesis, etc.
This function is most suitable when working with random effects models
fitted using \code{\link{lme}} or else.
}
\note{
Currently only 3 models are tidied (in join step) due to column name mismatch
}
\examples{
p_annotator(list(lme_model3, lme_model2, lme_model1))
}
