% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_biplot.R
\name{PCbiplot}
\alias{PCbiplot}
\title{It takes a prcomp object and returns a PCA biplot for selected variables}
\usage{
PCbiplot(PC, x = "PC1", y = "PC2")
}
\arguments{
\item{x}{Name of first principal component vector. i.e. x = "PC1"}

\item{y}{Name of second principal component vector. i.e. y = "PC2"}

\item{A}{`prcomp` object returned by prcomp()}
}
\value{
A ggplot2 object
}
\description{
This function is a convenience function for quick plotting of PCA biplot
}
\examples{

\dontrun{
PCbiplot(stats::prcomp(dfp %>%
                  group_by(genotype_id) %>%
                  summarise_if(is.numeric, mean, na.rm = T) %>%
                  select_at(which(map_lgl(., is.numeric))) %>%
                  select(-contains("rep")) %>%
                  scale())) +
  theme_bw() +
  theme(text = element_text(size = 15))
}

}
